Modularize NPO card
